+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+2.2.25
+------
+- Changed some method comments
+- Added Datamodel setIdentity method
+- Identitylinker refactoring
+
 2.2.24
 ------
 - Changed language core utility to intercept empty strings